refactor: modularize global settings
